Move test double files outside of Integration folder (#146)
