+++ tests/Akeneo/Data/Attribute/COLOR.json
