+++ tests/Akeneo/Data/Attribute/attachment.json
