+++ tests/Akeneo/Data/Attribute/material.json
