+++ tests/Akeneo/Data/Attribute/size.json
