+++ tests/Akeneo/Data/AttributeOption/size/l.json
