+++ tests/Akeneo/Data/AttributeOption/size/m.json
