+++ tests/Akeneo/Data/AttributeOption/size/s.json
