+++ tests/Akeneo/Data/Family/loudspeakers.json
