+++ tests/Akeneo/Data/FamilyVariant/Mugs/COLOR.json
