+++ tests/Akeneo/Data/FamilyVariant/accessories/accessories_size.json
