+++ tests/Akeneo/Data/Product/10627329.json
