+++ tests/Akeneo/Data/Product/127469.json
