+++ tests/Akeneo/Data/Product/1314976.json
