+++ tests/Akeneo/Data/Product/16466450.json
