+++ tests/Akeneo/Data/Product/BRAIDED_HAT_L.json
