+++ tests/Akeneo/Data/Product/BRAIDED_HAT_M.json
