+++ tests/Akeneo/Data/Product/BRAIDED_HAT_S.json
