+++ tests/Akeneo/Data/Product/MUG_DW.json
