+++ tests/Akeneo/Data/Product/MUG_SW.json
