+++ tests/Akeneo/Data/Product/no-name-product.json
