+++ tests/Akeneo/Data/Product/tshirt-new-m.json
