+++ tests/Akeneo/Data/ProductModel/MODEL_BRAIDED_HAT.json
