+++ tests/Akeneo/Data/media-files/1/3/9/e/139e9b32956237c28b5d9a36d00a34254075316f_media_11556.jpeg
