+++ tests/Akeneo/Data/media-files/4/5/b/4/45b4d93d07dfc795d884847d840c1876613992a7_braided_hat.jpg
