+++ tests/Akeneo/Data/media-files/4/f/e/0/4fe0ab031f723625958ccf684d72525837e8fae3_10627329_7290.jpg
