+++ tests/Akeneo/Data/media-files/8/e/2/d/8e2d8254e599201a0e6e91d0c512bc37c8780c21_1314976_5566.jpg
