+++ tests/Akeneo/Data/media-files/c/1/2/2/c1222e8291f3f5e4f4946baeb34f46b97a309218_c06324701.jpg
