+++ tests/Akeneo/TestDouble/FamilyApiMock.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Tests\Webgriffe\SyliusAkeneoPlugin\Integration\TestDouble;
+namespace Tests\Webgriffe\SyliusAkeneoPlugin\Akeneo\TestDouble;
 
-use Akeneo\Pim\ApiClient\Api\ProductModelApiInterface;
+use Akeneo\Pim\ApiClient\Api\FamilyApiInterface;
 use Akeneo\Pim\ApiClient\Pagination\PageInterface;
 use Akeneo\Pim\ApiClient\Pagination\ResourceCursorInterface;
 
-final class ProductModelApiMock implements ProductModelApiInterface
+final class FamilyApiMock implements FamilyApiInterface
 {
     public function create(string $code, array $data = []): int
     {
@@ -17,9 +17,7 @@ final class ProductModelApiMock implements ProductModelApiInterface
 
     public function get(string $code): array
     {
-        return ApiClientMock::jsonFileOrHttpNotFoundException(
-            __DIR__ . '/../DataFixtures/ApiClientMock/ProductModel/' . $code . '.json',
-        );
+        return $this->jsonDecodeOrNull(__DIR__ . '/../Data/Family/' . $code . '.json');
     }
 
     public function listPerPage(int $limit = 10, bool $withCount = false, array $queryParameters = []): PageInterface
@@ -42,8 +40,15 @@ final class ProductModelApiMock implements ProductModelApiInterface
         // TODO: Implement upsertList() method.
     }
 
-    public function delete(string $code): int
+    /**
+     * @return mixed|null
+     */
+    private function jsonDecodeOrNull(string $filename)
     {
-        // TODO: Implement delete() method.
+        if (file_exists($filename)) {
+            return json_decode(file_get_contents($filename), true, 512, \JSON_THROW_ON_ERROR);
+        }
+
+        return null;
     }
 }
