+++ tests/TestDouble/DateTimeBuilder.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Webgriffe\SyliusAkeneoPlugin\Integration\TestDouble;
+namespace Tests\Webgriffe\SyliusAkeneoPlugin\TestDouble;
 
 use Webgriffe\SyliusAkeneoPlugin\DateTimeBuilderInterface;
